requirements.txt
argon2-cffi==19.1.0
astroid==2.2.5
cffi==1.12.2
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
protobuf==3.7.1
pycparser==2.19
pylint==2.3.1
PyMySQL==0.9.3
six==1.12.0
SQLAlchemy==1.3.6
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1
